Refactored DynamicClass.cs
